+++ dotnet/Connectors/Options/DatabaseOptions.cs
@@ -1,0 +1,9 @@
+﻿namespace Connectors.Options
+{
+    public class DatabaseOptions
+    {
+        public string ReadConnectionString { get; set; } = string.Empty;
+
+        public string WriteConnectionString { get; set; } = string.Empty;
+    }
+}