+++ dotnet/Domain/Models/HousingLocation.cs
@@ -1,11 +1,13 @@
-﻿namespace Domain
+﻿namespace Domain.Models
 {
     public class HousingLocation
     {
+        // XXX? DDD "should NOT have 'Id's"
         public int? Id { get; set; } = null;
 
         public string? Name { get; set; } = null;
 
+        // TODO Group Citi/State
         public string? City { get; set; } = null;
 
         public string? State { get; set; } = null;